% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseplot.R
\name{plot_bandgaps}
\alias{plot_bandgaps}
\title{Plot band gaps}
\usage{
plot_bandgaps(materials, secondary.scale = "AVS", dev = FALSE)
}
\arguments{
\item{materials}{a subset of \code{\link{semiconductors}} dataframe}

\item{secondary.scale}{select reference electrode for the secondary y-axis
Default secondary scale is AVS (vacuum scale). Set this argument to empty
string to disable the secondary scale.}

\item{dev}{used for testing during development. Setting to TRUE enables extra
plot elements such as x-axis and grids to support development.}
}
\description{
Band energies and gaps of user-selectable semiconductors using R base graphics.
There is no practical way to return a plot object using base graphics, so
this function does not return anything, it just creates the plot.
It is probably well-suited for use in a Rmd or LaTeX document.
The generated plot requires the LaTeX packages siunitx and chemformula to
typeset the text labels.
}
\examples{
\dontrun{
plot_bandgaps(bandgaps::semiconductors \%>\%
   filter(class == "oxide") \%>\%
   filter(str_detect(formula, "Fe")),
   secondary.scale = "AVS")
plot_bandgaps(bandgaps::semiconductors \%>\%
   filter(class == "oxide") \%>\%
   filter(str_detect(formula, "O2")))
}
}
