% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{semiconductor_row}
\alias{semiconductor_row}
\title{Collect semiconductor properties into df row ready to add to dataset}
\usage{
semiconductor_row(formula, polymorph = "", xtal.size = "bulk",
  sctype = "", class = "", CB = NA, VB = NA, Eg = NA,
  transition = "", pH = NA, pH.ZPC = NA, ref, comment = "")
}
\arguments{
\item{formula}{mandatory. The compound's chemical formula.}

\item{polymorph}{optional. Polymorph name, e.g., anatase, brookite.}

\item{xtal.size}{optional, defaults to "bulk". Crystallite size.}

\item{sctype}{optional. Semiconductor type, i.e., n or p.}

\item{class}{optional, but recommended. Material class, e.g., oxide, sulfide, nitride}

\item{CB}{optional, but make sure that any two of: CB, VB or Eg are specified.}

\item{VB}{optional, but make sure that any two of: CB, VB or Eg are specified.}

\item{Eg}{optional, but make sure that any two of: CB, VB or Eg are specified.}

\item{transition}{optional, transition type: direct, indirect, etc.}

\item{pH}{optional, but recommended, allowed to be NA. The pH value at the CB/VB.}

\item{pH.ZPC}{optional. pH at which surface charge is neutral.}

\item{ref}{mandatory. BibTeX reference string.}

\item{comment}{optional.}
}
\value{
single-row dataframe ready to be incorporated into a semiconductors dataframe
}
\description{
IMPORTANT: always specify CB/VB potentials vs SHE.
May I suggest you use \code{\link[common]{as.SHE}} for this purpose.
}
\examples{
\dontrun{
semiconductor_row(formula="CdS", CB=common::as.SHE(-3.98, "AVS"), Eg=2.40, pH=2.00, ref="Xu2000")
}
}
