% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{semiconductors}
\alias{semiconductors}
\title{Band gaps and band edge potentials for semiconductors}
\format{A data frame with 27 rows and 14 variables:
\describe{
  \item{id}{label used for uniquely identifying each row}
  \item{formula}{The compound's chemical formula.}
  \item{polymorph}{Polymorph, e.g., anatase, brookite.}
  \item{xtal.size}{Crystallite size, e.g., bulk, micro, nano.}
  \item{sctype}{Semiconductor type, i.e., n or p.}
  \item{class}{Material class, e.g., oxide, sulfide, nitride.}
  \item{CB}{Conduction band edge potential/V(SHE).}
  \item{VB}{Valence band edge potential/V(SHE).}
  \item{Eg}{Band gap/V(SHE)}
  \item{transition}{Transition type: direct, indirect, etc.}
  \item{pH}{The pH at which the reported CB/VB occurs.}
  \item{pH.ZPC}{The pH at which the material's surface is neutral.}
  \item{ref}{Bibtex reference key. See citations in /data-raw/references.bib.}
  \item{comment}{Free-form comment.}
}}
\usage{
semiconductors
}
\description{
A dataset listing band gaps and band edge potentials
    for semiconductor materials. Sourced from primary literature sources.
}
\author{
Taha Ahmed
}
\keyword{datasets}
